edi_sale: move OrderMixin from edi_sale_ubl
